+++ CHANGELOG.md
@@ -2,6 +2,58 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.9.1-rc.5](https://github.com/inference-gateway/rust-sdk/compare/0.9.1-rc.4...0.9.1-rc.5) (2025-03-21)
+
+### ♻️ Improvements
+
+* Improve example for Tool-Use ([bc647da](https://github.com/inference-gateway/rust-sdk/commit/bc647dad7686aecd63a9fa38617dfe3a29d3bdb0))
+
+## [0.9.1-rc.4](https://github.com/inference-gateway/rust-sdk/compare/0.9.1-rc.3...0.9.1-rc.4) (2025-03-21)
+
+### ♻️ Improvements
+
+* Remove ToolFunctionResponse and update ToolCallResponse to use ChatCompletionMessageToolCallFunction ([e3d29bb](https://github.com/inference-gateway/rust-sdk/commit/e3d29bbd291d84708f4f37e0cdc338829b4cface))
+
+### 📚 Documentation
+
+* Enhance README with updated response types and improved logging ([651720b](https://github.com/inference-gateway/rust-sdk/commit/651720ba50cec350c92bce4f72a353c3d9189d9b))
+
+## [0.9.1-rc.3](https://github.com/inference-gateway/rust-sdk/compare/0.9.1-rc.2...0.9.1-rc.3) (2025-03-21)
+
+### ♻️ Improvements
+
+* Update API endpoints and logging for model listing ([96fec11](https://github.com/inference-gateway/rust-sdk/commit/96fec11450919679eafca4b46a436713831ccca9))
+
+## [0.9.1-rc.2](https://github.com/inference-gateway/rust-sdk/compare/0.9.1-rc.1...0.9.1-rc.2) (2025-03-21)
+
+### ♻️ Improvements
+
+* Update model response structures for clarity and consistency ([5ca2eb5](https://github.com/inference-gateway/rust-sdk/commit/5ca2eb5baad76b4ad3aaf405233036cd2e0b539c))
+
+### 🔧 Miscellaneous
+
+* Update dependencies for improved performance and security ([6079995](https://github.com/inference-gateway/rust-sdk/commit/607999577166554c32264327e20da2daa0b8b94e))
+
+## [0.9.1-rc.1](https://github.com/inference-gateway/rust-sdk/compare/0.9.0...0.9.1-rc.1) (2025-03-20)
+
+### ♻️ Improvements
+
+* Add streaming response structures for chat completion ([0dd7b56](https://github.com/inference-gateway/rust-sdk/commit/0dd7b5693f14d086acdacc7356b34caf73eb7035))
+* Download the OpenAPI spec ([7701059](https://github.com/inference-gateway/rust-sdk/commit/7701059e64ae6073e73803b1ddb4fbc93dc35417))
+* Remove Google provider from the Provider enum ([5a98051](https://github.com/inference-gateway/rust-sdk/commit/5a98051437c3cc6578b7de86ef937ae4736b7184))
+* Remove nullable fields from OpenAPI schema definitions ([461750f](https://github.com/inference-gateway/rust-sdk/commit/461750facb43b70a3112f867cb689711804ca560))
+* Rename default client method to new_default for clarity ([8735999](https://github.com/inference-gateway/rust-sdk/commit/87359996264e6d50a3ef078080794d3254ed9e06))
+* Rename GenerateRequest to CreateChatCompletionRequest and remove ssevents field ([52c4008](https://github.com/inference-gateway/rust-sdk/commit/52c40089ec6c59fce0c9748f8f1c985cd81e6fab))
+* Restructure response to be compatible with OpenAI ([88da800](https://github.com/inference-gateway/rust-sdk/commit/88da80062a23a272be04c54b2d81a667004ca066))
+* Start by implementing default and base_url ([1afdaac](https://github.com/inference-gateway/rust-sdk/commit/1afdaac83a7cccd8405db1eeddb387dc0822260e))
+* Update JSON response format to include new fields for chat completion ([38869a1](https://github.com/inference-gateway/rust-sdk/commit/38869a19293385f5c6660adb5c3b79afddf2b906))
+* Update JSON response structure with new fields for chat completion ([340c02c](https://github.com/inference-gateway/rust-sdk/commit/340c02c006aa397b0ce8e7b512b9bf9c6e244f8a))
+* Update tool call structures for chat completion and enhance argument parsing ([c867978](https://github.com/inference-gateway/rust-sdk/commit/c86797846676a266f477d906f1e9a9a2e12f41cd))
+
+### 🐛 Bug Fixes
+
+* Add output logging for version determination error in release workflow ([d49ba50](https://github.com/inference-gateway/rust-sdk/commit/d49ba50bd4111b28b722ae78657c6660909c0932))
+
 ## [0.9.0](https://github.com/inference-gateway/rust-sdk/compare/0.8.0...0.9.0) (2025-02-11)
 
 ### ✨ Features